Introduce the concept of 'parent deletion countering strategies'.
This is to handle cases where e.g. a cell is deleted on one side
and patched on the other side. Currently only inline-source is
marked as a countering strategy. This means if the cell patch
contains a change to source, and source has the inline-source
strategy, the cell will not be deleted. All other contents
of the patch will be applied as if onesided (i.e. no conflict
resolution needed, keeping the code simpler), and the
inline-source strategy will be applied to the source with
knowledge of the parent deletion passed to it such that it
can insert a <CELL DELETED> marker for the user to see and edit.